+++ internal/domain/estimate/api/handler.go
@@ -6,6 +6,7 @@ import (
 	"strconv"
 
 	"github.com/gin-gonic/gin"
+	"github.com/google/uuid"
 
 	"personal-finance/internal/domain/estimate"
 	"personal-finance/internal/domain/estimate/service"
@@ -25,6 +26,8 @@ func NewBalanceHandlers(r *gin.Engine, service service.Service) {
 	r.GET("/estimate", handler.FindEstimateByMonth())
 	r.POST("/estimate", handler.AddEstimate())
 	r.POST("/sub-estimate", handler.AddSubEstimate())
+	r.PUT("/estimate/:id", handler.UpdateEstimateAmount())
+	r.PUT("/sub-estimate/:id", handler.UpdateSubEstimateAmount())
 }
 
 func (h handler) FindEstimateByMonth() gin.HandlerFunc {
@@ -120,4 +123,68 @@ func (h handler) AddSubEstimate() gin.HandlerFunc {
 		return
 	}
 
+}
+
+func (h handler) UpdateEstimateAmount() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		userID, err := authentication.GetUserIDFromContext(c)
+		if err != nil {
+			c.JSON(http.StatusUnauthorized, err.Error())
+			return
+		}
+
+		idString := c.Param("id")
+		id, err := uuid.Parse(idString)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+
+		var input model.EstimateCategories
+		if err := c.BindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, err.Error())
+			return
+		}
+
+		updatedEstimate, err := h.service.UpdateEstimateAmount(c.Request.Context(), &id, input.Amount, userID)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+
+		c.JSON(http.StatusOK, updatedEstimate)
+		return
+	}
+}
+
+func (h handler) UpdateSubEstimateAmount() gin.HandlerFunc {
+	return func(c *gin.Context) {
+		userID, err := authentication.GetUserIDFromContext(c)
+		if err != nil {
+			c.JSON(http.StatusUnauthorized, err.Error())
+			return
+		}
+
+		idString := c.Param("id")
+		id, err := uuid.Parse(idString)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+
+		var input model.EstimateSubCategories
+		if err := c.BindJSON(&input); err != nil {
+			c.JSON(http.StatusBadRequest, err.Error())
+			return
+		}
+
+		updatedSubEstimate, err := h.service.UpdateSubEstimateAmount(c.Request.Context(), &id, input.Amount, userID)
+		if err != nil {
+			c.JSON(http.StatusInternalServerError, err.Error())
+			return
+		}
+
+		c.JSON(http.StatusOK, updatedSubEstimate)
+		return
+	}
 }